Major UI overhall of the content creator. I have rearranged the solution to make it a tidier folder structure, and refactored common components (e.g. the header) into actual components (rather than replicating the HTML and CSS on every page).
I have also tidied up the UI significantly in preperation for the pre-beta release of the app (for tech challenge) on the 18th July. I have added a welcome message, which explains the app is in pre-beta, ordered tiles on the home screen by availability (so links which work are at the top, and links which describe pages which will be added but are not yet in the system are further down and disabled with text to explain). This gives users a clear view of what features they can use today, and what they can expect in future (which will help us sell our app at tech challenge).